Prototype
The Gleaners truck drives around collecting cans which get added behind it. If the truck runs into one of these cans, the game is over.

I'm happy with the can, but the truck is a WIP.

Everything gravitates around the sphere/world.

Next, I'm going to add Cinemachine, spawn the cans around the surface of the world, and handle replaying and scoring.