Added sounds and fixed inconsistency at different framerates
Added 9 sounds for the game. These vary from a soundtrack to little noises that happen based on actions in the game (like picking up cans).

Fixed inconsistency at different framerates. The attractor (planet) was attracting attractables (pickups and truck) in Update. I have since changed that to FixedUpdate so that it happens consistently regardless of framerate.

With the consistency change, I changed the character movement from a rigidbody move to a transform translation as it also is more consistent.