Touch input WebGL
Added touch input as a secondary input for WebGL. Now, when the input is accepted, it will always check for CrossPlatformInput and then if there is any touch input.

This is so that the game will work for WebGL on mobile mostly, but the touch controls can also be used on any system which has a touch screen and an up-to-date browser.